(PC-21734)[PRO] fix: remove useParticipants to handle change via events instead
